requirements.txt
websockets
pillow
opencv-python
numpy
simpleaudio
matplotlib